+++ packages/astro-components-stencil/src/components/rux-input/rux-input.tsx
@@ -21,12 +21,12 @@ let id = 0
  * @part label - The input label when `label` prop is set
  */
 @Component({
-    tag: 'rux-input-field',
-    styleUrl: 'rux-input-field.scss',
+    tag: 'rux-input',
+    styleUrl: 'rux-input.scss',
     shadow: true,
 })
-export class RuxInputField implements FormFieldInterface {
-    @Element() el!: HTMLRuxInputFieldElement
+export class RuxInput implements FormFieldInterface {
+    @Element() el!: HTMLRuxInputElement
     inputId = `rux-input-${++id}`
 
     @State() hasLabelSlot = false