+++ packages/astro-components-stencil/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Breaking Changes
 
+#### Input Field
+
+-   Component has been renamed to `rux-input` for convenience.
+
 #### Checkbox
 
 -   `required` prop has been removed. Use Checkbox Group component instead to display invalid state.